fix: delete symlink if exist

+++ src/twitter/DataOrganizer.js
@@ -30,7 +30,9 @@ class DataOrganizer {
     }
 
     try {
-      await fs.unlink(this.latestPath)
+      if (fs.existsSync(this.latestPath)) {
+        await fs.unlink(this.latestPath)
+      }
       await fs.symlink(this.baseDir, this.latestPath);
       Logger.info(`✅ Create symlink: ${this.latestPath}`);
     } catch (error) {
